# BaseDatos/openapi-gateway.yaml
swagger: "2.0"
info:
  title: "basedatos"
  version: "1.0.0"
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  api_key:
    type: apiKey
    name: x-api-key
    in: header
security:
  - api_key: []

# Backend global: evita repetirlo en cada path
...
paths:
...
  /generar_secciones:
    options:
      operationId: generar_secciones_options
      security: []
      responses:
        "200":
          description: CORS preflight
    post:
      operationId: generar_secciones
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required: [centro_id, clase, asignatura, unidad, num_secciones]
            properties:
              centro_id:
                type: string
              clase:
                type: string
              asignatura:
                type: string
              unidad:
                type: integer
              num_secciones:
                type: integer
              consideraciones_adicionales:
                type: string
      responses:
        "200":
          description: OK
...